Feature/likesandcomments

+++ src/controllers/episode.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { ResponseHelper } from '../helpers';
+import { HttpError, ResponseHelper } from '../helpers';
 import {
   IEpisodeController,
   IEpisodeForm,
@@ -15,6 +15,12 @@ class EpisodeController implements IEpisodeController {
     this.createEpisode = this.createEpisode.bind(this);
     this.updateEpisode = this.updateEpisode.bind(this);
     this.deleteEpisode = this.deleteEpisode.bind(this);
+
+    this.likeEpisode = this.likeEpisode.bind(this);
+    this.getEpisodeLikes = this.getEpisodeLikes.bind(this);
+
+    this.createEpisodeComment = this.createEpisodeComment.bind(this);
+    this.getEpisodeComments = this.getEpisodeComments.bind(this);
   }
 
   async getAllEpisodes(req: Request, res: Response) {
@@ -30,6 +36,11 @@ class EpisodeController implements IEpisodeController {
 
   async getEpisodeById(req: Request, res: Response) {
     const { episode_id } = req.params;
+
+    if (!parseInt(episode_id)) {
+      throw new HttpError(HttpStatusCode.MethodNotAllowed, "Method not allowed");
+    }
+
     const episode = await this.episodeService.getEpisodeById(
       parseInt(episode_id),
     );
@@ -47,6 +58,11 @@ class EpisodeController implements IEpisodeController {
     const limit = parseInt(req.query.limit as string) || 10;
 
     const { creator_id } = req.params;
+
+    if (!parseInt(creator_id)) {
+      throw new HttpError(HttpStatusCode.MethodNotAllowed, "Method not allowed");
+    }
+
     const episodes = await this.episodeService.getEpisodesByCreatorId(parseInt(creator_id), page, limit);
 
     return ResponseHelper.responseSuccess(
@@ -98,6 +114,10 @@ class EpisodeController implements IEpisodeController {
   async updateEpisode(req: Request, res: Response) {
     const { episode_id } = req.params;
 
+    if (!parseInt(episode_id)) {
+      throw new HttpError(HttpStatusCode.MethodNotAllowed, "Method not allowed");
+    }
+
     const episodeData: IEpisodeForm = req.body;
 
     const creator_id = res.locals.user.user_id;
@@ -132,6 +152,10 @@ class EpisodeController implements IEpisodeController {
   async deleteEpisode(req: Request, res: Response) {
     const { episode_id } = req.params;
 
+    if (!parseInt(episode_id)) {
+      throw new HttpError(HttpStatusCode.MethodNotAllowed, "Method not allowed");
+    }
+
     const episode = await this.episodeService.deleteEpisode(
       parseInt(episode_id),
     );
@@ -141,6 +165,71 @@ class EpisodeController implements IEpisodeController {
       HttpStatusCode.Ok,
       'Episode deleted successfully',
       episode,
+    );
+  }
+
+  async likeEpisode (req: Request, res: Response) {
+    const {
+      episode_id
+    } = req.body;
+
+    const isLiked = await this.episodeService.likeEpisode(episode_id, parseInt(res.locals.id));
+
+    return ResponseHelper.responseSuccess(
+      res,
+      HttpStatusCode.Ok,
+      `Episode ${isLiked ? 'Liked': 'Unliked'} successfully`
+    );
+  }
+
+  async getEpisodeLikes (req: Request, res: Response) {
+    const { episode_id } = req.params;
+
+    if (!parseInt(episode_id)) {
+      throw new HttpError(HttpStatusCode.MethodNotAllowed, "Method not allowed");
+    }
+
+    const count = await this.episodeService.getEpisodeLikes(parseInt(episode_id));
+
+    return ResponseHelper.responseSuccess(
+      res,
+      HttpStatusCode.Ok,
+      "Operation successful",
+      {
+        count
+      }
+    );
+  }
+
+  async createEpisodeComment (req: Request, res: Response) {
+    const { episode_id, username, comment_text } = req.body;
+
+    console.log(episode_id, comment_text);
+
+    const comment = await this.episodeService.createEpisodeComment(episode_id, parseInt(res.locals.id), username, comment_text);
+
+    return ResponseHelper.responseSuccess(
+      res,
+      HttpStatusCode.Created,
+      'Comment created successfully',
+      comment,
+    );
+  }
+
+  async getEpisodeComments (req: Request, res: Response) {
+    const { episode_id } = req.params;
+
+    if (!parseInt(episode_id)) {
+      throw new HttpError(HttpStatusCode.MethodNotAllowed, "Method not allowed");
+    }
+
+    const comments = await this.episodeService.getEpisodeComments(parseInt(episode_id));
+
+    return ResponseHelper.responseSuccess(
+      res,
+      HttpStatusCode.Ok,
+      "Operation successful",
+      comments
     );
   }
 }
